requirements.txt
asgiref==3.10.0
Django==5.2.7
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
django-allauth==0.56.0
environs[django]==14.5.0
marshmallow<3.21.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0    
django-cors-headers==4.9.0
dj-rest-auth[with_social]==6.0.0
drf-spectacular==0.28.0